requirements.txt
django-htcpcp-tea==0.3.1
django-mailer==2.0.1
django-phonenumber-field==5.0.0
django-tables2==2.3.1
django-rest-polymorphic==0.1.9
django-bootstrap-datepicker-plus==3.0.5
django-colorfield==0.3.2
